chore: seperate file uploader and tx confirmation modal from script deployer
